// portal/proto/isuxportal/error_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/error.proto (package isuxportal.proto, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file isuxportal/error.proto.
 */
export declare const file_isuxportal_error: GenFile;

/**
 * @generated from message isuxportal.proto.Error
 */
export declare type Error = Message<"isuxportal.proto.Error"> & {
  /**
   * @generated from field: int32 code = 1;
   */
  code: number;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: string human_message = 3;
   */
  humanMessage: string;

  /**
   * @generated from field: repeated string human_descriptions = 4;
   */
  humanDescriptions: string[];

  /**
   * @generated from field: isuxportal.proto.Error.DebugInfo debug_info = 16;
   */
  debugInfo?: Error_DebugInfo;
};

/**
 * Describes the message isuxportal.proto.Error.
 * Use `create(ErrorSchema)` to create a new message.
 */
export declare const ErrorSchema: GenMessage<Error>;

/**
 * @generated from message isuxportal.proto.Error.DebugInfo
 */
export declare type Error_DebugInfo = Message<"isuxportal.proto.Error.DebugInfo"> & {
  /**
   * @generated from field: string exception = 1;
   */
  exception: string;

  /**
   * @generated from field: repeated string trace = 2;
   */
  trace: string[];

  /**
   * @generated from field: repeated string application_trace = 3;
   */
  applicationTrace: string[];

  /**
   * @generated from field: repeated string framework_trace = 4;
   */
  frameworkTrace: string[];
};

/**
 * Describes the message isuxportal.proto.Error.DebugInfo.
 * Use `create(Error_DebugInfoSchema)` to create a new message.
 */
export declare const Error_DebugInfoSchema: GenMessage<Error_DebugInfo>;