// portal/proto/isuxportal/services/audience/team_list_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/services/audience/team_list.proto (package isuxportal.proto.services.audience, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file isuxportal/services/audience/team_list.proto.
 */
export declare const file_isuxportal_services_audience_team_list: GenFile;

/**
 * @generated from message isuxportal.proto.services.audience.ListTeamsRequest
 */
export declare type ListTeamsRequest = Message<"isuxportal.proto.services.audience.ListTeamsRequest"> & {
};

/**
 * Describes the message isuxportal.proto.services.audience.ListTeamsRequest.
 * Use `create(ListTeamsRequestSchema)` to create a new message.
 */
export declare const ListTeamsRequestSchema: GenMessage<ListTeamsRequest>;

/**
 * @generated from message isuxportal.proto.services.audience.ListTeamsResponse
 */
export declare type ListTeamsResponse = Message<"isuxportal.proto.services.audience.ListTeamsResponse"> & {
  /**
   * @generated from field: repeated isuxportal.proto.services.audience.ListTeamsResponse.TeamListItem teams = 1;
   */
  teams: ListTeamsResponse_TeamListItem[];
};

/**
 * Describes the message isuxportal.proto.services.audience.ListTeamsResponse.
 * Use `create(ListTeamsResponseSchema)` to create a new message.
 */
export declare const ListTeamsResponseSchema: GenMessage<ListTeamsResponse>;

/**
 * @generated from message isuxportal.proto.services.audience.ListTeamsResponse.TeamListItem
 */
export declare type ListTeamsResponse_TeamListItem = Message<"isuxportal.proto.services.audience.ListTeamsResponse.TeamListItem"> & {
  /**
   * @generated from field: int64 team_id = 1;
   */
  teamId: bigint;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: repeated string member_names = 3;
   */
  memberNames: string[];

  /**
   * @generated from field: bool final_participation = 4;
   */
  finalParticipation: boolean;

  /**
   * @generated from field: bool is_student = 5;
   */
  isStudent: boolean;
};

/**
 * Describes the message isuxportal.proto.services.audience.ListTeamsResponse.TeamListItem.
 * Use `create(ListTeamsResponse_TeamListItemSchema)` to create a new message.
 */
export declare const ListTeamsResponse_TeamListItemSchema: GenMessage<ListTeamsResponse_TeamListItem>;