// portal/proto/isuxportal/services/registration/env_check_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/services/registration/env_check.proto (package isuxportal.proto.services.registration, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";
import type { EnvCheckStatus } from "../../resources/env_check_pb";

/**
 * Describes the file isuxportal/services/registration/env_check.proto.
 */
export declare const file_isuxportal_services_registration_env_check: GenFile;

/**
 * @generated from message isuxportal.proto.services.registration.GetEnvCheckInformationResponse
 */
export declare type GetEnvCheckInformationResponse = Message<"isuxportal.proto.services.registration.GetEnvCheckInformationResponse"> & {
  /**
   * @generated from field: string template = 1;
   */
  template: string;

  /**
   * @generated from field: isuxportal.proto.resources.EnvCheckStatus status = 2;
   */
  status: EnvCheckStatus;

  /**
   * @generated from field: string instance_ip = 3;
   */
  instanceIp: string;
};

/**
 * Describes the message isuxportal.proto.services.registration.GetEnvCheckInformationResponse.
 * Use `create(GetEnvCheckInformationResponseSchema)` to create a new message.
 */
export declare const GetEnvCheckInformationResponseSchema: GenMessage<GetEnvCheckInformationResponse>;