// portal/proto/isuxportal/services/registration/activate_coupon_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/services/registration/activate_coupon.proto (package isuxportal.proto.services.registration, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file isuxportal/services/registration/activate_coupon.proto.
 */
export declare const file_isuxportal_services_registration_activate_coupon: GenFile;

/**
 * @generated from message isuxportal.proto.services.registration.ActivateCouponRequest
 */
export declare type ActivateCouponRequest = Message<"isuxportal.proto.services.registration.ActivateCouponRequest"> & {
  /**
   * @generated from field: int64 team_id = 1;
   */
  teamId: bigint;
};

/**
 * Describes the message isuxportal.proto.services.registration.ActivateCouponRequest.
 * Use `create(ActivateCouponRequestSchema)` to create a new message.
 */
export declare const ActivateCouponRequestSchema: GenMessage<ActivateCouponRequest>;

/**
 * @generated from message isuxportal.proto.services.registration.ActivateCouponResponse
 */
export declare type ActivateCouponResponse = Message<"isuxportal.proto.services.registration.ActivateCouponResponse"> & {
};

/**
 * Describes the message isuxportal.proto.services.registration.ActivateCouponResponse.
 * Use `create(ActivateCouponResponseSchema)` to create a new message.
 */
export declare const ActivateCouponResponseSchema: GenMessage<ActivateCouponResponse>;