// portal/proto/isuxportal/services/contestant/notifications_pb.d.ts
export declare type ListNotificationsResponse = Message<"isuxportal.proto.services.contestant.ListNotificationsResponse"> & {
  /**
   * @generated from field: int64 last_answered_clarification_id = 1;
   */
  lastAnsweredClarificationId: bigint;

  /**
   * @generated from field: repeated isuxportal.proto.resources.Notification notifications = 2;
   */
  notifications: Notification[];
};

/**
 * Describes the message isuxportal.proto.services.contestant.ListNotificationsResponse.
 * Use `create(ListNotificationsResponseSchema)` to create a new message.
 */
export declare const ListNotificationsResponseSchema: GenMessage<ListNotificationsResponse>;

/**
 * @generated from message isuxportal.proto.services.contestant.SubscribeNotificationRequest
 */
export declare type SubscribeNotificationRequest = Message<"isuxportal.proto.services.contestant.SubscribeNotificationRequest"> & {
  /**
   * @generated from field: string endpoint = 1;
   */
  endpoint: string;

  /**
   * @generated from field: string p256dh = 2;
   */
  p256dh: string;

  /**
   * @generated from field: string auth = 3;
   */
  auth: string;
};

/**
 * Describes the message isuxportal.proto.services.contestant.SubscribeNotificationRequest.
 * Use `create(SubscribeNotificationRequestSchema)` to create a new message.
 */
export declare const SubscribeNotificationRequestSchema: GenMessage<SubscribeNotificationRequest>;

/**
 * @generated from message isuxportal.proto.services.contestant.SubscribeNotificationResponse
 */
export declare type SubscribeNotificationResponse = Message<"isuxportal.proto.services.contestant.SubscribeNotificationResponse"> & {
};

/**
 * Describes the message isuxportal.proto.services.contestant.SubscribeNotificationResponse.
 * Use `create(SubscribeNotificationResponseSchema)` to create a new message.
 */
export declare const SubscribeNotificationResponseSchema: GenMessage<SubscribeNotificationResponse>;

/**
 * @generated from message isuxportal.proto.services.contestant.UnsubscribeNotificationRequest
 */
export declare type UnsubscribeNotificationRequest = Message<"isuxportal.proto.services.contestant.UnsubscribeNotificationRequest"> & {
  /**
   * @generated from field: string endpoint = 1;
   */
  endpoint: string;
};

/**
 * Describes the message isuxportal.proto.services.contestant.UnsubscribeNotificationRequest.
 * Use `create(UnsubscribeNotificationRequestSchema)` to create a new message.
 */
export declare const UnsubscribeNotificationRequestSchema: GenMessage<UnsubscribeNotificationRequest>;

/**
 * @generated from message isuxportal.proto.services.contestant.UnsubscribeNotificationResponse
 */
export declare type UnsubscribeNotificationResponse = Message<"isuxportal.proto.services.contestant.UnsubscribeNotificationResponse"> & {
};

/**
 * Describes the message isuxportal.proto.services.contestant.UnsubscribeNotificationResponse.
 * Use `create(UnsubscribeNotificationResponseSchema)` to create a new message.
 */
export declare const UnsubscribeNotificationResponseSchema: GenMessage<UnsubscribeNotificationResponse>;