// portal/proto/isuxportal/services/contestant/dashboard_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/services/contestant/dashboard.proto (package isuxportal.proto.services.contestant, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";
import type { LeaderboardItem } from "../../resources/leaderboard_pb";

/**
 * Describes the file isuxportal/services/contestant/dashboard.proto.
 */
export declare const file_isuxportal_services_contestant_dashboard: GenFile;

/**
 * @generated from message isuxportal.proto.services.contestant.DashboardQuery
 */
export declare type DashboardQuery = Message<"isuxportal.proto.services.contestant.DashboardQuery"> & {
};

/**
 * Describes the message isuxportal.proto.services.contestant.DashboardQuery.
 * Use `create(DashboardQuerySchema)` to create a new message.
 */
export declare const DashboardQuerySchema: GenMessage<DashboardQuery>;

/**
 * @generated from message isuxportal.proto.services.contestant.DashboardResponse
 */
export declare type DashboardResponse = Message<"isuxportal.proto.services.contestant.DashboardResponse"> & {
  /**
   * @generated from field: isuxportal.proto.resources.LeaderboardItem leaderboard_item = 1;
   */
  leaderboardItem?: LeaderboardItem;
};

/**
 * Describes the message isuxportal.proto.services.contestant.DashboardResponse.
 * Use `create(DashboardResponseSchema)` to create a new message.
 */
export declare const DashboardResponseSchema: GenMessage<DashboardResponse>;